+++ tests/dummy-robot.js
@@ -14,46 +14,14 @@
 
 "use strict";
 
-var Log = require('log');
+var Logger = require('./dummy-logger.js');
 
-function Logger(enabled) {
-  this.enabled = enabled;
-  this.log = new Log('debug');
-
-  this.error = function(msg) {
-    if (!this.enabled) {
-      return;
-    }
-    this.log.error(msg);
-  };
-
-  this.warning = function(msg) {
-    if (!this.enabled) {
-      return;
-    }
-    this.log.warning(msg);
-  };
-
-  this.info = function(msg) {
-    if (!this.enabled) {
-      return;
-    }
-    this.log.info(msg);
-  };
-
-  this.debug = function(msg) {
-    if (!this.enabled) {
-      return;
-    }
-    this.log.debug(msg);
-  };
-}
-
-function Robot(name, adapter, enable_logging) {
+function Robot(name, adapter, enable_logging, robot_name) {
   this.logger = new Logger(enable_logging);
   this.name = name;
   this.commands = [];
   this.adapter = adapter;
+  this.robot_name = robot_name;
 
   this.messageRoom = function(recipient, data) {
     return;